requirements.txt
# Core dependencies
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
psutil==5.9.6
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.4
Pillow==10.1.0

# Windows-specific monitoring (required for production)
pywin32==306 ; platform_system == "Windows"
wmi==1.5.1 ; platform_system == "Windows"

# Additional utilities
click==8.1.7
colorama==0.4.6
